Дорабатывает изменение избранного
